src/requirements.txt
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
fastapi==0.78.0
flower==1.1.0
h11==0.13.0
humanize==4.2.3
idna==3.3
kombu==5.2.4
packaging==21.3
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
pydantic==1.9.1
pyparsing==3.0.9
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
redis==4.3.4
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tornado==6.2
typing_extensions==4.3.0
uvicorn==0.18.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
